polling

+++ server/score_evaluation/usecase/score_evaluation_usecase.py
@@ -7,8 +7,10 @@ class ScoreEvaluationUsecase:
 
     def execute(self):
         eval = self.evaluation_message_repository_interface.fetch_message()
-        eval_app = ScoreEvaluationApplication(eval)
-        eval = eval_app.evaluate()
-        if eval.status == "S":
-            self.evaluation_message_repository_interface.delete_message(eval)
+        if eval.repository_url != "":
+            eval_app = ScoreEvaluationApplication(eval)
+            eval = eval_app.evaluate()
+            if eval.status == "S":
+                self.evaluation_message_repository_interface.delete_message(eval)
         return eval
+
